Make RDB files point to a (new) AOF segment. Both loading and saving works.
AOF rewrites don't work right now. This will be fixed later. A new RDB
version is introduced. Old append only files and RDBs are handled
as expected. The AOF is only loaded directly when an RDB doesn't exist or if
it's in an old format (and AOF is enabled, of course).